% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{aggregate_peaks}
\alias{aggregate_peaks}
\title{Identify peaks by aggregating differentially methylated regions}
\usage{
aggregate_peaks(dm_results, regions = regions_gatc_drosophila_dm6)
}
\arguments{
\item{dm_results}{data frame of differential methylation results obtained from [edgeR_results]}

\item{regions}{data frame of GATC regions. Default is GATC regions from Drosophila melanogaster - dm6.}
}
\value{
data frame of peaks. columns are chromosome name, start, end, and width of peak, peak identifier, n of dm regions within the peak, gap in bp to the next peak, and peak rank (based on p value).
}
\description{
`aggregate_peaks` aggregates differentially methylated GATC regions into peaks. These peaks represent the region that the gene of interest bound in.
}
