% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{aggregate_peaks_new}
\alias{aggregate_peaks_new}
\title{New: identify peaks from de results}
\usage{
aggregate_peaks_new(dm_results, regions = regions_gatc_drosophila_dm6)
}
\arguments{
\item{dm_results}{data.frame as outputted from `edgeR_results()`}

\item{regions}{regions df. default is the provided data -
`regions_gatc_drosophila_dm6`}
}
\value{
data.frame of peaks
}
\description{
New: identify peaks from de results
}
