% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{aggregate_peaks_old}
\alias{aggregate_peaks_old}
\title{OLD:Identify peaks by aggregating differentially methylated regions}
\usage{
aggregate_peaks_old(dm_results)
}
\arguments{
\item{dm_results}{data.frame of differential methylation results obtained from [edgeR_results]}
}
\value{
data.frame of peaks.
Columns are
* seqnames,
* start,
* end,
* width,
* consec_dm (peak identifier),
* n_regions_dm (n of dm regions within the peak),
* ave_logFC (average logFC calculated across regions),
* ave_pVal (average adjusted P Value calculated across the regions),
* rank_p (based on average p value).
* gap (gap in bp to the next peak),
* number (peak number in order of position)
}
\description{
`aggregate_peaks_old` aggregates differentially methylated GATC regions into peaks. These peaks represent the region that the gene of interest bound in.
}
