% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{annotate_genes_new}
\alias{annotate_genes_new}
\title{New: annotation of peaks and genes}
\usage{
annotate_genes_new(peaks, genes, regions = regions_gatc_drosophila_dm6)
}
\arguments{
\item{peaks}{peaks df as outputted from [aggregate_peaks()]}

\item{genes}{genes df as outputted from [get_biomart_genes()]}

\item{regions}{regions df, default is [regions_gatc_drosophila_dm6]}
}
\value{
list of 3 dfs: closest - every peak with it's closest gene,
combo - every peak with list of 5 closest genes,
all - all genes matching to each peak and all information
}
\description{
New: annotation of peaks and genes
}
