% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_results.R
\name{corr_scatter}
\alias{corr_scatter}
\title{Plotting correlation of specific samples}
\usage{
corr_scatter(df, sample_1, sample_2, method = "spearman")
}
\arguments{
\item{df}{data frame as outputted from [process_bams()]}

\item{sample_1}{name of the BAM file within df}

\item{sample_2}{name of the BAM file within df}

\item{method}{correlation method. Default is the non-parametric spearman's. Non-parametric methods are recommended as data does not reliably meet the requirements for parametric analysis.}
}
\value{
A scatterplot of the two selected samples, showing their counts at each region, overlaid with the correlation results.
}
\description{
`corr_scatter` plots a scatterplot with correlation results from two selected samples, in order to visualise their similarity/difference.
* uses the non-parametric "spearman's" correlation.
* for more information, see: [ggpubr::ggscatter()]
}
\details{
The correlation between Dam_1 and Fusion_1 can be expected to reach ~ 0.7, whereas the correlation between Dam_1 & Dam_3 or Fusion_1 & Fusion_2 would be expected to be closer to ~0.9
}
