% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_meth_edgeR.R
\name{edgeR_results_plot}
\alias{edgeR_results_plot}
\title{Plot differential testing results}
\usage{
edgeR_results_plot(dge, p.value = 0.05, lfc = 1)
}
\arguments{
\item{dge}{DGEList object as outputted from [edgeR_set_up()]}

\item{p.value}{The p value threshold for minimum significance. Default is 0.05}

\item{lfc}{The minimum log fold change for significant results. Default is 1}
}
\value{
An MA style scatter plot with average logCPM on x-axis, average logFC on y-axis, with dots coloured by significance
}
\description{
`edgeR_results_plot` provides an MA style plot for differential methylation results, allowing for a visualisation of the logFC, P values, and spread of -1,0,1 results.
* for further details, see [edgeR::plotSmear()]
}
\examples{
counts.df <- random_counts()
dge <- edgeR_set_up(counts.df)

edgeR_results_plot(dge, p.value = 0.05, lfc = 1)
}
