% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_meth_edgeR.R
\name{edgeR_set_up}
\alias{edgeR_set_up}
\title{Create DGE object for differential testing}
\usage{
edgeR_set_up(
  counts.df,
  max.width = 10000,
  lib.size = NULL,
  keep_a = 0.5,
  keep_b = 3
)
}
\arguments{
\item{counts.df}{A data.frame generated from [process_bams]. Ensure that the samples are ordered by (Dam_1.bam, Fusion_1.bam, Dam_2.bam, Fusion_2.bam, ...)}

\item{max.width}{Remove large regions, default is width of 10,000. We recommend this value as the Dam can methylate GATC sites up to 5kb away from the binding site, generating a total width of 10 kb.}

\item{lib.size}{Library size for each sample is calculated as the sum across all rows for that sample unless otherwise specified}

\item{keep_a}{Filtering parameter, minimum counts per million (cpm) of each sample. Recommend leaving at default of 0.5}

\item{keep_b}{Filtering parameter, minimum number of samples to meet the criteria of keep_a in order to retain the region in the downstream analysis. Default is 3 (assuming 6 samples)}
}
\value{
An object of class `DGEList`. Refer to [edgeR::?`DGEListClass`] for details
}
\description{
`edgeR_set_up()` sets up the edgeR analysis for visualisation of the samples [edgeR_plot_mds()], and then for identifying differentially methylated regions [edgeR_results()].
}
\examples{
counts.df <- random_counts()

edgeR_set_up(counts.df)
}
