% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatc_track.R
\name{gatc_region_fn}
\alias{gatc_region_fn}
\title{Extract GATC regions}
\usage{
gatc_region_fn(BSgenome_object = NULL, path_to_fasta = NULL)
}
\arguments{
\item{BSgenome_object}{A BSgenome package.}

\item{path_to_fasta}{An ensembl}
}
\value{
A list object composed of two data.frames
* GATC regions for use in analysis
* GATC sites, can be used in plotting of results
}
\description{
`gatc_region_fn` identifies and extracts the GATC sites and regions from a BSgenome object or a fasta file.
}
