% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{gene_annotate}
\alias{gene_annotate}
\title{Annotating peaks to their closest gene}
\usage{
gene_annotate(peaks, genes)
}
\arguments{
\item{peaks}{data.frame of peaks as outputted from [aggregate_peaks]}

\item{genes}{data frame of gene information as outputted from [get_biomart_genes]}
}
\value{
data.frame of annotated peaks and genes in a horrible messy format I don't like to look at.
}
\description{
`gene_annotate` takes both the peaks and genes as input and returns the paired results.
}
