% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{gene_annotate_organised}
\alias{gene_annotate_organised}
\title{OLD: Tabular display of peak statistical information and their closest genes}
\usage{
gene_annotate_organised(annotated_peaks)
}
\arguments{
\item{annotated_peaks}{data.frame of annotated peaks [gene_annotate()]}
}
\value{
organised data frame of results with statistical information about each peak, alongside a list of the closest genes.
* contains information about the peak and gene location (for the closest gene), and provides a list of other nearby genes and their distance to the peak.
}
\description{
`gene_annotate_organised` simplifies the output from `gene_annotate` to a clean format.
}
