% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_de_lfc.R
\name{geom_de.res.lfc}
\alias{geom_de.res.lfc}
\title{Plotting de results with lfc}
\usage{
geom_de.res.lfc(de_results.df, plot.space = 0.1, plot.height = 0.1)
}
\arguments{
\item{de_results.df}{A data.frame of differential testing results as outputted from `edgeR_results()`}

\item{plot.space}{Specify gap to next plot, default is 0.1}

\item{plot.height}{Specify overall height of plot, default is 0.1}
}
\value{
ggplot_add object
}
\description{
`geom_de.res.lfc` is a ggplot layer that visualises the de_results and logFC across a given region.
* regions are coloured by de result: 1, 0, -1, NA (grey for NA)
* cannot be plotted by itself, must be added to an existing plot - see examples.
}
\examples{
counts.df <- random_counts()
dge <- edgeR_set_up(counts.df)
de_results <- random_edgeR_results()
plot_counts_all_bams(counts.df,
                     seqnames = "chr2L",
                     start_region = 1,
                     end_region = 40000,
                     n_col = 1) +
  geom_de.res.lfc(de_results)
}
