% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_de.R
\name{geom_de.res}
\alias{geom_de.res}
\title{Plotting de results}
\usage{
geom_de.res(de_results.df = NULL, plot.space = 0.1, plot.height = 0.3)
}
\arguments{
\item{de_results.df}{results df plus add_de: `add_de(edgeR_results())`}

\item{plot.space}{gap to next plot}

\item{plot.height}{height of plot}
}
\description{
Plotting de results
}
