% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gatc_track.R
\name{geom_gatc}
\alias{geom_gatc}
\title{Plot for a GATC track}
\usage{
geom_gatc(
  gatc.df = NULL,
  gatc.color = "red",
  gatc.size = 5,
  plot.space = 0.1,
  plot.height = 0.1
)
}
\arguments{
\item{gatc.df}{df of positions of GATC sites - can be made from GATC region file}

\item{gatc.color}{default is red}

\item{gatc.size}{size of line}

\item{plot.space}{gap to next plot}

\item{plot.height}{size of plot}
}
\description{
Plot for a GATC track
}
