% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peak.R
\name{geom_peak.new}
\alias{geom_peak.new}
\title{Plotting peaks}
\usage{
geom_peak.new(
  peak.df,
  peak.color = "black",
  peak.size = 5,
  plot.space = 0.1,
  plot.height = 0.1
)
}
\arguments{
\item{peak.df}{data frame of peak locations}

\item{peak.color}{can set the colour if you want}

\item{peak.size}{default is 5}

\item{plot.space}{gap to next plot - default 0.1}

\item{plot.height}{height of plot - leave it at default}

\item{bed.file}{bed input of peaks}
}
\value{
plot
}
\description{
`geom_peak.new` adds the peak location to an existing ggplot2 object
}
\examples{
counts.df <- process_bams(path_to_bams = system.file("extdata", package = "Damsel"), regions = regions_gatc_drosophila_dm6, cores = 2)
counts.df <- counts.df[,c(1:6,7,10,8,11,9,12)]
counts.df
dge <- edgeR_set_up(counts.df)
de_results <- edgeR_results(dge, p.value = 0.05, lfc = 1)
peaks <- aggregate_peaks(de_results, regions = regions_gatc_drosophila_dm6)
peaks
plot_counts_all_bams(counts_df, seqnames = "chr2L", start_region = 1, end_region = 40000, n_col = 1) +
  geom_peak.new(peaks)
# The plots can be layered -------------------------------------------------------------------------------------
}
