% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_counts.R
\name{geom_regions.counts}
\alias{geom_regions.counts}
\title{GGplot addable counts graph for all samples}
\usage{
geom_regions.counts(
  region.df = NULL,
  n_col = 1,
  region.color = "black",
  plot.space = 0.1,
  plot.height = 1
)
}
\arguments{
\item{region.df}{df of counts}

\item{n_col}{n of columns to facet by}

\item{region.color}{colour}

\item{plot.space}{gap to next plot}

\item{plot.height}{height of plot}
}
\description{
GGplot addable counts graph for all samples
}
