% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_lfc.R
\name{geom_regions.lfc}
\alias{geom_regions.lfc}
\title{Plot lfc results across the regions}
\usage{
geom_regions.lfc(
  region.df,
  region.color = "black",
  plot.space = 0.1,
  plot.height = 1
)
}
\arguments{
\item{region.df}{results df plus add_de: `add_de(edgeR_results())`}

\item{region.color}{leave as default}

\item{plot.space}{gap to next plot}

\item{plot.height}{height of plot}
}
\description{
Plot lfc results across the regions
}
