% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{get_biomart_genes}
\alias{get_biomart_genes}
\title{Get gene information from Ensembl using biomaRt}
\usage{
get_biomart_genes(
  species,
  version = 109,
  regions = regions_gatc_drosophila_dm6
)
}
\arguments{
\item{species}{species of interest. Format is first letter of genus, followed by full name of species, followed by gene_ensembl. For example: Drosophila melanogaster is dmelanogaster_gene_ensembl}

\item{version}{Ensembl version of genome. Default is 109 (the most recent update to the Drosophila melanogaster dm6 genome)}

\item{regions}{data frame of GATC regions. Default is GATC regions from Drosophila melanogaster - dm6.}
}
\value{
data.frame of information about the genes. Columns include (I don't know what order right now): chromosome, start, end, width, strand, gene name, ensembl_gene_id, ensembl_transcript_id, n_overlapping_gatc_regions
}
\description{
`get_biomart_genes` accesses the Ensembl database via [biomaRt::useEnsembl() and biomaRt::getBM()] to obtain the location of genes from the selected species.
Also identifies the number of GATC regions matching to each gene.
}
