% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper_fns.R
\name{GetRegion_hack}
\alias{GetRegion_hack}
\title{Get the regions}
\usage{
GetRegion_hack(df, chr, start, end = NULL)
}
\arguments{
\item{df}{df used by plot fn}

\item{chr}{seqnames}

\item{start}{start of plotting region}

\item{end}{end of plotting region}
}
\description{
helper fn
}
