% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_de.R
\name{ggplot_add.de.res}
\alias{ggplot_add.de.res}
\title{Constructor for de results plot}
\usage{
ggplot_add.de.res(object, plot, object_name)
}
\arguments{
\item{object}{de results}

\item{plot}{plot being added to}

\item{object_name}{de results}
}
\description{
Constructor for de results plot
}
