% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gatc_track.R
\name{ggplot_add.gatc}
\alias{ggplot_add.gatc}
\title{Constructor for GATC track plot}
\usage{
ggplot_add.gatc(object, plot, object_name)
}
\arguments{
\item{object}{GATC track}

\item{plot}{plot above}

\item{object_name}{GATC track}
}
\description{
Constructor for GATC track plot
}
