% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peak.R
\name{ggplot_add.peak.new}
\alias{ggplot_add.peak.new}
\title{how to make the above plot}
\usage{
ggplot_add.peak.new(object, plot, object_name)
}
\arguments{
\item{object}{this is what goes in - peak.df or bed.file}

\item{plot}{this is the plot you're adding to}

\item{object_name}{"peak.new}
}
\description{
code for the `geom_peak.new` plot
}
