% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_counts.R
\name{ggplot_add.regions.counts}
\alias{ggplot_add.regions.counts}
\title{Constructor for ggplot addable}
\usage{
\method{ggplot_add}{regions.counts}(object, plot, object_name)
}
\arguments{
\item{object}{regions.counts}

\item{plot}{plot being added to}

\item{object_name}{regions.counts}
}
\description{
Constructor for ggplot addable
}
