% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_lfc.R
\name{ggplot_add.regions.lfc}
\alias{ggplot_add.regions.lfc}
\title{Constructor for plotting lfc results across regions}
\usage{
\method{ggplot_add}{regions.lfc}(object, plot, object_name)
}
\arguments{
\item{object}{thing to plot}

\item{plot}{plot being added to}

\item{object_name}{thing to plot}
}
\description{
Constructor for plotting lfc results across regions
}
