% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ontology.R
\name{goseq_fn}
\alias{goseq_fn}
\title{Gene ontology analysis}
\usage{
goseq_fn(regions, genes, peaks)
}
\arguments{
\item{regions}{data frame of regions between GATC sites - load data}

\item{genes}{data frame of gene data}

\item{peaks}{data frame of annotated peak data}
}
\value{
4 objects
 * List of top 10 over-represented GO terms across the 3 GO categories
 * Plot of goodness of fit of model
 * Plot of sample data
 * Plot of sample data without bias correction (should be messy)
}
\description{
`goseq_fn` identifies the top 10 over-represented GO terms from the peak data, correcting for the number of GATC regions matching to each gene.
}
