% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ideogram.R
\name{hack_adjustZoom}
\alias{hack_adjustZoom}
\title{adjust the zoom of the highlight region}
\usage{
hack_adjustZoom(obj, plot, zoom.region, zoom.offset, color, fill, size, alpha)
}
\arguments{
\item{obj}{as above}

\item{plot}{as above}

\item{zoom.region}{as above}

\item{zoom.offset}{as above}

\item{color}{as above}

\item{fill}{as above}

\item{size}{as above}

\item{alpha}{as above}
}
\description{
adjust the zoom of the highlight region
}
