% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ideogram.R
\name{hack_applyTheme}
\alias{hack_applyTheme}
\title{Theme for ideogram}
\usage{
hack_applyTheme(plot, xlabel, subchr, aspect.ratio)
}
\arguments{
\item{plot}{as above}

\item{xlabel}{as above}

\item{subchr}{as above}

\item{aspect.ratio}{as above}
}
\description{
Theme for ideogram
}
