% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ideogram.R
\name{hack_Ideogram}
\alias{hack_Ideogram}
\title{Plot ideogram}
\usage{
hack_Ideogram(
  obj,
  subchr = NULL,
  which = NULL,
  xlabel = FALSE,
  cytobands = TRUE,
  color = "red",
  fill = "red",
  alpha = 0.7,
  zoom.region = NULL,
  zoom.offset = 0.2,
  size = 1,
  aspect.ratio = 1/20,
  ...,
  genome
)
}
\arguments{
\item{obj}{as [ggbio::Ideogram()]}

\item{subchr}{chromosome to use}

\item{which}{granges}

\item{xlabel}{position}

\item{cytobands}{TRUE/FALSE}

\item{color}{highlight colour}

\item{fill}{highlight fill colour}

\item{alpha}{transparency of highlight}

\item{zoom.region}{default is NULL}

\item{zoom.offset}{see ggbio}

\item{size}{ooh is this what I want to adjust?}

\item{aspect.ratio}{how big it looks}

\item{...}{other args passed on}

\item{genome}{if don't have an object give a genome name}
}
\value{
plot
}
\description{
Plot ideogram
}
