% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ideogram.R
\name{hack_selectChromosome}
\alias{hack_selectChromosome}
\title{helper for hack_Ideogram - select chromosome}
\usage{
hack_selectChromosome(obj, subchr)
}
\arguments{
\item{obj}{see `hack_Ideogram`}

\item{subchr}{chromsoome to use}
}
\description{
helper for hack_Ideogram - select chromosome
}
