% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{new_peaks_fn}
\alias{new_peaks_fn}
\title{New peaks function}
\usage{
new_peaks_fn(dm_results, gap_size = 150)
}
\arguments{
\item{dm_results}{results from differential testing}

\item{gap_size}{maximum gap in bp between dm regions}
}
\value{
df of peaks ranked by p-value
}
\description{
New peaks function
}
\examples{
set.seed(123)
counts.df <- random_counts()
dm_results <- random_edgeR_results()
peaks <- new_peaks_fn(dm_results)
peaks
}
