% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ontology.R
\name{plot_gene_ontology}
\alias{plot_gene_ontology}
\title{Plot gene ontology results}
\usage{
plot_gene_ontology(
  signif_results,
  plot_type = c("bar", "dot"),
  bar_x = c("gene_ratio", "gene_count", "-log10FDR")
)
}
\arguments{
\item{signif_results}{results as outputted from goseq_fn()$signif_results. Selects the top 20 GO terms as default}

\item{plot_type}{Plot results as a bar or a dot plot. Bar is default method}

\item{bar_x}{Select x axis for bar plot method= c(gene_ratio, gene_count, -log10FDR). Default is gene_ratio}
}
\value{
A ggplot2 object
}
\description{
Plot gene ontology results
}
\examples{
set.seed(123)
example_regions <- random_regions()
peaks <- aggregate_peaks(random_edgeR_results())
genes <- get_biomart_genes(
    species = "dmelanogaster_gene_ensembl",
    version = 109,
    regions = example_regions
)
annotation <- annotate_genes(peaks, genes, example_regions)$all

ontology <- goseq_fn(annotation, genes, example_regions)$signif_results
plot_gene_ontology(ontology, plot_type = "bar", bar_x = "gene_ratio")
plot_gene_ontology(ontology, plot_type = "dot")

}
