% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bams.R
\name{process_bams}
\alias{process_bams}
\title{Initial processing of BAM files}
\usage{
process_bams(path_to_bams, regions = regions_gatc_drosophila_dm6, cores = 2)
}
\arguments{
\item{path_to_bams}{Character string input vector. Should identify directory containing the BAM files.}

\item{regions}{Data frame of GATC regions. Default is GATC regions from Drosophila melanogaster - dm6.}

\item{cores}{Specify computer cores to be used to parallelise function and speed output. Default is to use all available cores. If computer is being used for multiple tasks at once, reccommend reducing the number of cores. Number of available cores can be checked using [parallel::detectCores()]}
}
\value{
A data frame containing the GATC region information in the form in the columns: seqnames (chromosome), start, end, width, and strand. The count information for the BAM files is in the subsequent columns, named by the name of the BAM file.
* If necessary, at this stage please rearrange the BAM file columns so they are ordered in the following way: Dam_1, Fusion_1, Dam_2, Fusion_2 etc
* The DamID data captures the ~75bp region extending from each GATC site, so although regions are of differing widths, there is a null to minimal length bias present on the data, and does not require length correction.
}
\description{
`process_bams()` obtains the counts for the regions between GATC sites, from BAM files specified in the path.
}
