% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{random_counts}
\alias{random_counts}
\title{Create example counts}
\usage{
random_counts(size = 50)
}
\arguments{
\item{size}{number of rows to create}
}
\value{
example data.frame of counts similar to `process_bams()`
}
\description{
Create example counts
}
\examples{
head(random_counts(size = 50))
}
