% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{random_edgeR_results}
\alias{random_edgeR_results}
\title{Create example edgeR results}
\usage{
random_edgeR_results(size = 50)
}
\arguments{
\item{size}{number of rows to create}
}
\value{
example data.frame of edgeR results, output similar to `edgeR_results()`
}
\description{
Create example edgeR results
}
\examples{
head(random_edgeR_results(size = 50))
}
