% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{random_regions}
\alias{random_regions}
\title{Create example regions}
\usage{
random_regions(size = 50)
}
\arguments{
\item{size}{number of rows to create}
}
\value{
example data.frame with output similar to `gatc_track()$regions`
}
\description{
Create example regions
}
\examples{
random_regions(size = 50)
}
