% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper_fns.R
\name{theme_peak_hack}
\alias{theme_peak_hack}
\title{theme for narrow plots}
\usage{
theme_peak_hack(margin.len, x.range)
}
\arguments{
\item{margin.len}{how big a gap}

\item{x.range}{the plotting region}
}
\description{
Theme used
}
