% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{add_de}
\alias{add_de}
\title{Add de_results to full region df}
\usage{
add_de(de_results, regions = regions_gatc_drosophila_dm6)
}
\arguments{
\item{de_results}{as outputted from [edgeR_results()]}

\item{regions}{data.frame of regions, default is regions_gatc_drosophila_dm6}
}
\value{
full data.frame of regions with added information about the de results;
* de - 1,0,-1,NA ;
* logFC: 0 if de is NA ;
* adjust.p: 1 if de is NA :
* meth_status: Upreg, No_sig, Downreg, Not_included
}
\description{
Used within aggregate_peaks to add in the regions that were excluded from edgeR analysis for low counts
Is also required for some plotting fns
}
\examples{
counts.df <- process_bams(path_to_bams = system.file("extdata", package = "Damsel"), regions = regions_gatc_drosophila_dm6, cores = 2)
counts.df <- counts.df[,c(1:6,7,10,8,11,9,12)]
counts.df
dge <- edgeR_set_up(counts.df)
dge
de_results <- edgeR_results(dge, p.value = 0.05, lfc = 1)
de_results
add_de(de_results, regions = regions_gatc_drosophila_dm6)
}
