% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{aggregate_peaks}
\alias{aggregate_peaks}
\title{New: identify peaks from dm results}
\usage{
aggregate_peaks(dm_results)
}
\arguments{
\item{dm_results}{A data.frame of differential testing results for each GATC region as outputted from `edgeR_results()`}
}
\value{
A data.frame of peaks. Columns are as follows: peak_id (Unique peak identifier, used internally - PS indicates a single peak, PM indicates the peak was combined), seqnames, start, end, width, strand, rank_p, ave_logFC, ave-pVal, multiple_peaks (number of peaks, NA if 1), n_regions_dm, n_regions_not_dm
}
\description{
`aggregate_peaks()` aggregates upregulated regions as outputted from `edgeR_results()`. These peaks represent the region that the transcription factor of interest bound in.

The average logFC and adjusted p-value is calculated for each peak, and peaks are ranked by their new p-value. Peaks with a gap between them of <= 150 bp are combined into 1 peak, accounting for many of the small regions having little data.
}
\examples{
dm_results <- random_edgeR_results()

aggregate_peaks(dm_results)
}
