% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{aggregate_peaks}
\alias{aggregate_peaks}
\title{Identify peaks by aggregating differentially methylated regions}
\usage{
aggregate_peaks(dm_results, regions = regions_gatc_drosophila_dm6)
}
\arguments{
\item{dm_results}{data frame of differential methylation results obtained from [edgeR_results]}

\item{regions}{data frame of GATC regions. Default is GATC regions from Drosophila melanogaster - dm6.}
}
\value{
data frame of peaks.
Columns are
* seqnames,
* start,
* end,
* width,
* consec_dm (peak identifier),
* n_regions_dm (n of dm regions within the peak),
* ave_logFC (average logFC calculated across regions),
* ave_pVal (average adjusted P Value calculated across the regions),
* rank_p (based on average p value).
* gap (gap in bp to the next peak),
* number (peak number in order of position)
}
\description{
`aggregate_peaks` aggregates differentially methylated GATC regions into peaks. These peaks represent the region that the gene of interest bound in.
}
\examples{
path_to_bams <- system.file("extdata", package = "Damsel")
counts.df <- process_bams(path_to_bams,
                          regions = regions_gatc_drosophila_dm6,
                          cores = 2)
counts.df <- counts.df[,c(1:6,7,10,8,11,9,12)]
dge <- edgeR_set_up(counts.df)
de_results <- edgeR_results(dge, p.value = 0.05, lfc = 1)

peaks <- aggregate_peaks(de_results, regions = regions_gatc_drosophila_dm6)
head(peaks)
}
