% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{collateGenes}
\alias{collateGenes}
\title{New: Get list of genes}
\usage{
collateGenes(genes, regions, org.Db = NULL, version = NULL)
}
\arguments{
\item{genes}{A Txdb object, path to file, or a species for accessing biomart.}

\item{regions}{GATC region file}

\item{org.Db}{Required if using a Txdb object so to access gene names}

\item{version}{Required for using biomart}
}
\value{
GRanges object of genes and available supplementary information - specifically the TSS, and number of GATC regions overlapping the gene.
}
\description{
takes a Txdb object, path to a gff file, or a species (biomart) and returns a GRanges of genes.
}
\examples{
library(TxDb.Dmelanogaster.UCSC.dm6.ensGene)
library(org.Dm.eg.db)
set.seed(123)
example_regions <- random_regions()
txdb <- TxDb.Dmelanogaster.UCSC.dm6.ensGene
genes <- collateGenes(genes=txdb, regions=example_regions, org.Db=org.Dm.eg.db)
head(genes)

}
