% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Damsel-package.R
\docType{package}
\name{Damsel-package}
\alias{Damsel}
\alias{Damsel-package}
\title{Damsel: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
