% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_meth_edgeR.R
\name{edgeR_plot_mds}
\alias{edgeR_plot_mds}
\title{Multidimensional scaling plot for samples}
\usage{
edgeR_plot_mds(dge)
}
\arguments{
\item{dge}{as outputted from [edgeR_set_up()]}
}
\value{
refer to [edgeR::plotMDS()] for details
}
\description{
`edgeR_plot_mds` visualises the difference between samples.
* expect control (Dam-only) samples to cluster together and for Fusion samples to cluster together
}
\examples{
path_to_bams <- system.file("extdata", package = "Damsel")
counts.df <- process_bams(path_to_bams,
                          regions = regions_gatc_drosophila_dm6,
                          cores = 2)
counts.df <- counts.df[,c(1:6,7,10,8,11,9,12)]
dge <- edgeR_set_up(counts.df)

edgeR_plot_mds(dge)
}
