% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_meth_edgeR.R
\name{edgeR_results}
\alias{edgeR_results}
\title{Differential methylation analysis}
\usage{
edgeR_results(dge, p.value = 0.05, lfc = 1)
}
\arguments{
\item{dge}{as outputted from [edgeR_set_up()]}

\item{p.value}{p value threshold for minimum significance. Default is 0.05}

\item{lfc}{minimum log fold change for significant results. Default is 1}
}
\value{
data frame of differential methylation results. Columns are as follows; rownames(Region position), logFC (log fold change), logCPM (log counts per million), F (F statistic used to identify significance), PValue, adjustedP (tbd), dm (result: -1,0,1)
}
\description{
`edgeR_results` calculates the differential methylation results, identifying which GATC regions have been enriched in the Fusion samples relative to the controls.
Refer to the following pages for further details:
* [edgeR::glmQLFit()]
* [edgeR::glmGQLFTest()]
* [edgeR::decideTestsDGE()]
}
\examples{
counts.df <- process_bams(path_to_bams = system.file("extdata", package = "Damsel"), regions = regions_gatc_drosophila_dm6, cores = 2)
counts.df <- counts.df[,c(1:6,7,10,8,11,9,12)]
counts.df
dge <- edgeR_set_up(counts.df)
dge
edgeR_results(dge, p.value = 0.05, lfc = 1)
}
