% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatc_track.R
\name{gatc_region_fn}
\alias{gatc_region_fn}
\title{Extract GATC regions}
\usage{
gatc_region_fn(object)
}
\arguments{
\item{object}{A BSgenome package OR the path to a FASTA file.}
}
\value{
A `list` object composed of two `data.frames`.

The first; "regions" provides the GATC regions used in downstream analysis. The second, "sites", provides the positions of the GATC sites, and is used in plotting the results.
}
\description{
`gatc_region_fn` identifies and extracts the GATC sites and regions from a BSgenome object or a fasta file.
}
\examples{
if (require("BSgenome.Dmelanogaster.UCSC.dm6")) {
    gatc <- gatc_region_fn(BSgenome.Dmelanogaster.UCSC.dm6::BSgenome.Dmelanogaster.UCSC.dm6)

    head(gatc$regions)

    head(gatc$sites)
}
}
