% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{gene_annotate_organised}
\alias{gene_annotate_organised}
\title{Tabular display of peak statistical information and their closest genes}
\usage{
gene_annotate_organised(annotated_peaks)
}
\arguments{
\item{annotated_peaks}{data.frame of annotated messy peaks I don't like from [gene_annotate()]}
}
\value{
organised data frame of results with statistical information about each peak, alongside a list of the closest genes.
}
\description{
`gene_annotate_organised` simplifies the output from `gene_annotate` to a clean format. meaning I probably want it all as 1 fn
}
