% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_de_lfc.R
\name{geom_de.res.lfc}
\alias{geom_de.res.lfc}
\title{Plotting de results with lfc}
\usage{
geom_de.res.lfc(de_results.df, plot.space = 0.1, plot.height = 0.1)
}
\arguments{
\item{de_results.df}{results as outputted from `add_de(edgeR_results())`}

\item{plot.space}{space to next plot}

\item{plot.height}{height of plot}
}
\value{
ggplot_add object
}
\description{
`geom_de.res.lfc` is a ggplot layer that visualises the de_results and logFC across a given region.
* regions are coloured by de result: 1, 0, -1, NA (grey for NA)
* cannot be plotted by itself, must be added to an existing plot - see examples.
}
\examples{
path_to_bams <- system.file("extdata", package = "Damsel")
counts.df <- process_bams(path_to_bams,
                          regions = regions_gatc_drosophila_dm6,
                          cores = 2)
counts.df <- counts.df[,c(1:6,7,10,8,11,9,12)]
dge <- edgeR_set_up(counts.df)
de_results <- edgeR_results(dge, p.value = 0.05, lfc = 1)
de_results <- add_de(de_results, regions = regions_gatc_drosophila_dm6)
plot_counts_all_bams(counts.df,
                     seqnames = "chr2L",
                     start_region = 1,
                     end_region = 40000,
                     n_col = 1) +
  geom_de.res.lfc(de_results)
# The plots can be layered ------------------------------------------------
}
