% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gatc_track.R
\name{geom_gatc}
\alias{geom_gatc}
\title{Plot for a GATC track}
\usage{
geom_gatc(
  gatc.df = NULL,
  gatc.color = "red",
  gatc.size = 5,
  plot.space = 0.2,
  plot.height = 0.05
)
}
\arguments{
\item{gatc.df}{A data.frame of positions of GATC sites - can be made from GATC region file or from `gatc_track()$sites`}

\item{gatc.color}{Specify colour of lines, default is red}

\item{gatc.size}{Specify size of the line, default is 5}

\item{plot.space}{Specify gap to next plot, default is 0.2}

\item{plot.height}{Specify overall height of the plot, default is 0.05}
}
\value{
ggplot_add object.
}
\description{
`geom_gatc` is a ggplot layer that visualises the positions of GATC sites across a given region.
* cannot be plotted by itself, must be added to an existing ggplot object - see examples.
}
\examples{
path_to_bams <- system.file("extdata", package = "Damsel")
counts.df <- process_bams(path_to_bams,
                          regions = regions_gatc_drosophila_dm6,
                          cores = 2)
counts.df <- counts.df[,c(1:6,7,10,8,11,9,12)]
gatc_sites <- dplyr::mutate(regions_gatc_drosophila_dm6,
                            seqnames = paste0("chr", seqnames),
                            start = start - 3, end = start + 4)

plot_counts_all_bams(counts.df,
                     seqnames = "chr2L",
                     start_region = 1,
                     end_region = 40000,
                     n_col = 1) +
  geom_gatc(gatc_sites)
# The plots can be layered -------------------------------------------------
}
