% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gatc_track.R
\name{geom_gatc}
\alias{geom_gatc}
\title{Plot for a GATC track}
\usage{
geom_gatc(
  gatc.df = NULL,
  gatc.color = "red",
  gatc.size = 5,
  plot.space = 0.2,
  plot.height = 0.05
)
}
\arguments{
\item{gatc.df}{df of positions of GATC sites - can be made from GATC region file}

\item{gatc.color}{default is red}

\item{gatc.size}{size of line}

\item{plot.space}{gap to next plot}

\item{plot.height}{size of plot}
}
\value{
plot
}
\description{
Plot for a GATC track
}
\examples{
path_to_bams <- system.file("extdata", package = "Damsel")
counts.df <- process_bams(path_to_bams,
                          regions = regions_gatc_drosophila_dm6,
                          cores = 2)
counts.df <- counts.df[,c(1:6,7,10,8,11,9,12)]
gatc_sites <- dplyr::mutate(regions_gatc_drosophila_dm6,
                            seqnames = paste0("chr", seqnames),
                            start = start - 3, end = start + 4)

plot_counts_all_bams(counts.df,
                     seqnames = "chr2L",
                     start_region = 1,
                     end_region = 40000,
                     n_col = 1) +
  geom_gatc(gatc_sites)
# The plots can be layered -------------------------------------------------
}
