% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes.R
\name{geom_genes.me}
\alias{geom_genes.me}
\title{Plotting genes}
\usage{
geom_genes.me(
  df,
  txdb,
  gene_limits = NULL,
  plot.space = 0.1,
  plot.height = 0.3
)
}
\arguments{
\item{df}{A data.frame of genes as outputted from `get_biomart_genes`}

\item{gene_limits}{Set the height of the transcripts generated by `ggbio::autoplot()`. Default is NULL.
* If the gene is disproportionately large for the plot space, we recommend reducing the size with gene_limits = c(0,2)
* If there are a large amount of transcripts present, we recommend increasing the overall limit, example: gene_limits = c(0,11)}

\item{plot.space}{Specify gap to next plot, default is 0.1}

\item{plot.height}{Specify overall height of plot, default is 0.3}
}
\value{
ggplot layer
}
\description{
`geom_genes.me` is a ggplot layer that visualises the positions of genes across a given region.
* cannot be plotted by itself, must be added to an existing ggplot object - see examples.
}
\examples{
if (require("TxDb.Dmelanogaster.UCSC.dm6.ensGene")) {
  set.seed(123)
  example_regions <- random_regions()
  counts.df <- random_counts()
  genes <- get_biomart_genes(species = "dmelanogaster_gene_ensembl",
                             version = 109,
                             regions = example_regions)

  txdb <- TxDb.Dmelanogaster.UCSC.dm6.ensGene::TxDb.Dmelanogaster.UCSC.dm6.ensGene

  plot_counts_all_bams(counts.df,
                       seqnames = "chr2L",
                       start_region = 1,
                       end_region = 40000,
                       n_col = 1) +
    geom_genes.me(genes, txdb)
}
}
