% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes.R
\name{geom_genes.me}
\alias{geom_genes.me}
\title{Plotting genes}
\usage{
geom_genes.me(df, plot.space = 0.1, plot.height = 0.3)
}
\arguments{
\item{df}{genes as outputted from `get_biomart_genes`}

\item{plot.space}{gap to next plot}

\item{plot.height}{height of plot}
}
\value{
ggplot layer
}
\description{
Plotting genes
}
\examples{
path_to_bams <- system.file("extdata", package = "Damsel")
counts.df <- process_bams(path_to_bams,
                          regions = regions_gatc_drosophila_dm6,
                          cores = 2)
counts.df <- counts.df[,c(1:6,7,10,8,11,9,12)]
genes <- get_biomart_genes(species = "dmelanogaster_gene_ensembl",
                           version = 109,
                           regions = regions_gatc_drosophila_dm6)

plot_counts_all_bams(counts.df,
                     seqnames = "chr2L",
                     start_region = 1,
                     end_region = 40000,
                     n_col = 1) +
  geom_genes.me(genes)

}
