% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes.R
\name{geom_genes.me}
\alias{geom_genes.me}
\title{Plotting genes}
\usage{
geom_genes.me(df, plot.space = 0.1, plot.height = 0.3)
}
\arguments{
\item{df}{genes as outputted from `get_biomart_genes`}

\item{plot.space}{gap to next plot}

\item{plot.height}{height of plot}
}
\value{
ggplot layer
}
\description{
Plotting genes
}
\examples{
counts.df <- readRDS(test_path("fixtures", "test_counts_df.rds"))
genes <- readRDS(test_path("fixtures", "test_genes.rds"))

head(counts.df)
head(genes)

plot_counts_all_bams(counts.df,
                     seqnames = "chr2L",
                     start_region = 1,
                     end_region = 40000,
                     n_col = 1) +
  geom_genes.me(genes)

}
