% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peak.R
\name{geom_peak.new}
\alias{geom_peak.new}
\title{Plotting peaks}
\usage{
geom_peak.new(
  peaks.df = NULL,
  peak.label = FALSE,
  peak.color = "black",
  peak.size = 5,
  plot.space = 0.1,
  plot.height = 0.05
)
}
\arguments{
\item{peaks.df}{A data.frame of peaks as outputted from `aggregate_peaks()`}

\item{peak.label}{Specify whether peak_id labels should be added to the plot. Default is FALSE}

\item{peak.color}{Specify colour of peak. Default is black}

\item{peak.size}{Specify size of rectangle. Default is 5}

\item{plot.space}{Specify gap to next plot. Recommend leaving to the default: 0.1}

\item{plot.height}{Specify overall height of plot. Recommend leaving to the default: 0.05}
}
\value{
A `ggplot_add` object.
}
\description{
`geom_peak.new` is a ggplot layer that visualises the positions of peaks across a given region.
* cannot be plotted by itself, must be added to an existing ggplot object - see examples.
}
\examples{
set.seed(123)
counts.df <- random_counts()
dm_results <- random_edgeR_results()
peaks <- aggregate_peaks(dm_results)
plot_counts_all_bams(counts.df,
    seqnames = "chr2L",
    start_region = 1,
    end_region = 40000,
    n_col = 1
) +
    geom_peak.new(peaks)

plot_counts_all_bams(counts.df,
    seqnames = "chr2L",
    start_region = 1,
    end_region = 40000,
    n_col = 1
) +
    geom_peak.new(peaks, peak.label = TRUE)
# The plots can be layered -------------------------------------------------
}
