% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peak.R
\name{geom_peak.new}
\alias{geom_peak.new}
\title{Plotting peaks}
\usage{
geom_peak.new(
  bed.file = NULL,
  peak.df = NULL,
  peak.color = "black",
  peak.size = 5,
  plot.space = 0.1,
  plot.height = 0.1
)
}
\arguments{
\item{bed.file}{bed input of peaks}

\item{peak.df}{data frame of peak locations}

\item{peak.color}{can set the colour if you want}

\item{peak.size}{default is 5}

\item{plot.space}{gap to next plot - default 0.1}

\item{plot.height}{height of plot - leave it at default}
}
\description{
`geom_peak.new` adds the peak location to an existing ggplot2 object
}
