% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ontology.R
\name{goseq_fn}
\alias{goseq_fn}
\title{Gene ontology analysis}
\usage{
goseq_fn(annotation, genes, regions, extend_by = 2000, bias = NULL)
}
\arguments{
\item{annotation}{A data.frame of annotated genes and peaks as `annotate_peaks()$all`}

\item{genes}{A data.frame of gene data as outputted from `get_biomart_genes()`}

\item{regions}{A data.frame of GATC regions.}

\item{extend_by}{A number to extend the start and end of the genes. We recommend leaving to the default of 2000 bp.
* This is done to incorporate the acceptable distance of a peak to a gene.
* This also allows for consistency across significant and non-significant genes}

\item{bias}{alternatively, the bias can be input by itself.}
}
\value{
4 objects
 * List of top 10 over-represented GO terms across the 3 GO categories
 * Plot of goodness of fit of model
 * Plot of sample data
 * Plot of sample data without bias correction (should be messy)
}
\description{
`goseq_fn` identifies the top 10 over-represented GO terms from the peak data, correcting for the number of GATC regions matching to each gene.
}
\examples{
set.seed(123)
example_regions <- random_regions()
peaks <- aggregate_peaks(random_edgeR_results())
genes <- get_biomart_genes(
    species = "dmelanogaster_gene_ensembl",
    version = 109,
    regions = example_regions
)
annotation <- annotate_genes(peaks, genes, example_regions)$all

goseq_fn(annotation, genes, example_regions)
}
