% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_counts.R
\name{plot_counts_all_bams}
\alias{plot_counts_all_bams}
\title{Plot for just counts for all samples across a given region}
\usage{
plot_counts_all_bams(
  df,
  seqnames,
  start_region = NULL,
  end_region = NULL,
  n_col = 3
)
}
\arguments{
\item{df}{df of counts as outputted from [process_bams()]}

\item{seqnames}{chromosome of interest}

\item{start_region}{start}

\item{end_region}{end}

\item{n_col}{n_columns to facet the graph by - default is 3}
}
\value{
ggplot
}
\description{
Plot for just counts for all samples across a given region
}
\examples{
path_to_bams <- system.file("extdata", package = "Damsel")
counts.df <- process_bams(path_to_bams,
                          regions = regions_gatc_drosophila_dm6,
                          cores = 2)
counts.df <- counts.df[,c(1:6,7,10,8,11,9,12)]
plot_counts_all_bams(counts_df,
                     seqnames = "chr2L",
                     start_region = 1,
                     end_region = 40000,
                     n_col = 3)
# Can use this plot to layer other plots, just set n_col = 1 and start layering --------------------------------------
dge <- edgeR_set_up(counts.df)
de_results <- edgeR_results(dge, p.value = 0.05, lfc = 1)
de_results <- add_de(de_results, regions = regions_gatc_drosophila_dm6)
plot_counts_all_bams(counts_df,
                     seqnames = "chr2L",
                     start_region = 1,
                     end_region = 40000,
                     n_col = 1) +
  geom_de.res(de_results)
}
