% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_counts.R
\name{plot_counts_all_bams}
\alias{plot_counts_all_bams}
\title{Plot for just counts for all samples across a given region}
\usage{
plot_counts_all_bams(df, seqnames, start_region, end_region, n_col = 3)
}
\arguments{
\item{df}{df of counts as outputted from [process_bams()]}

\item{seqnames}{chromosome of interest}

\item{start_region}{start}

\item{end_region}{end}

\item{n_col}{n_columns to facet the graph by - default is 3}
}
\description{
Plot for just counts for all samples across a given region
}
