% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrap.R
\name{plot_wrap}
\alias{plot_wrap}
\title{Wrapper function for plotting}
\usage{
plot_wrap(
  peak_id = NULL,
  gene_id = NULL,
  seqnames = NULL,
  start_region = NULL,
  end_region = NULL,
  counts.df = NULL,
  dm_results.df = NULL,
  peaks.df = NULL,
  genes.df = NULL,
  txdb = NULL,
  gatc_sites.df = NULL,
  extend_by = 250,
  ...
)
}
\arguments{
\item{peak_id}{A character vector of a peak identifier(s) if wish to plot in peak centric manner. Default is NULL.}

\item{gene_id}{A character vector of ensembl gene id(s) if want to plot in gene centric manner. Default is NULL.}

\item{seqnames}{A chromosome. Default is NULL.}

\item{start_region}{A number providing the start of region to plot. Default is NULL.}

\item{end_region}{A number providing the end of region to plot. Default is NULL.}

\item{counts.df}{A data.frame of counts as from [process_bams()]. Default is NULL.}

\item{dm_results.df}{A data.frame of dm results as from [edgeR_results()]. Default is NULL.}

\item{peaks.df}{A data.frame of peaks as from [aggregate_peaks()]. Default is NULL.}

\item{genes.df}{A data.frame of genes as from [get_biomart_genes()]. Default is NULL.}

\item{txdb}{A TxDb object as from a TxDb package. Default is NULL.}

\item{gatc_sites.df}{A data.frame of gatc sites as from [gatc_track()$sites]. Default is NULL.}

\item{extend_by}{extend region. Default is 250}

\item{...}{arguments passed to geom_genes.me. Allows for adjusting of the plot appearance via gene_limits and plot.height if necessary.
* Default for gene_limits is NULL. If the gene is disproportionately large for the plot space, we recommend reducing the size with gene_limits = c(0,2)}
}
\value{
A `ggplot` object - or list of plots if provided multiple peaks/genes
}
\description{
`plot_wrap` plots all the available plots at once
}
