% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrap.R
\name{plot_wrap}
\alias{plot_wrap}
\title{Wrapper function for plotting}
\usage{
plot_wrap(
  peak_id = NULL,
  gene_id = NULL,
  seqnames = NULL,
  start_region = NULL,
  end_region = NULL,
  counts = NULL,
  de_results = NULL,
  peaks = NULL,
  genes = NULL,
  txdb = NULL,
  gatc_sites = NULL,
  extend_by = 250,
  ...
)
}
\arguments{
\item{peak_id}{peak identifier if wish to plot in peak centric manner}

\item{gene_id}{ensembl gene id if want to plot in gene centric manner}

\item{seqnames}{chromosome}

\item{start_region}{start of region to plot}

\item{end_region}{end of region to plot}

\item{counts}{counts.df from [process_bams()]}

\item{de_results}{edgeR results}

\item{peaks}{peaks data.frame}

\item{genes}{genes data.frame [get_biomart_genes()]}

\item{gatc_sites}{gatc sites}

\item{extend_by}{extend region, default is 250}

\item{...}{arguments passed to geom_genes.me. Allows for adjusting of the plot appearance via gene_limits and plot.height if necessary.
* Default for gene_limits is NULL. If the gene is disproportionately large for the plot space, we recommend reducing the size with gene_limits = c(0,2)}
}
\value{
ggplot object - or list of plots if provided multiple peaks/genes
}
\description{
`plot_wrap` plots all the available plots at once
}
